Fix many bugs

+++ jobs/states/delete.js
@@ -4,11 +4,10 @@ let State = models.states;
 module.exports = function(connection, done) {
   connection.createChannel(function(err, ch) {
     console.log(err);
-    var ex = 'kanban.main';
-    var queue = 'kanban.state.delete';
-    ch.assertExchange(ex, 'fanout', { durable: false });
-    ch.assertQueue(queue, { exclusive: false }, function(err, q) {
-      ch.bindQueue(q.queue, ex, queue);
+    var ex = 'chiepherd.main';
+    ch.assertExchange(ex, 'topic');
+    ch.assertQueue('kanban.state.delete', { exclusive: false }, function(err, q) {
+      ch.bindQueue(q.queue, ex, "kanban.state.delete")
 
       ch.consume(q.queue, function(msg) {
         // LOG
@@ -16,14 +15,57 @@ module.exports = function(connection, done) {
         let json = JSON.parse(msg.content.toString());
 
         // Create state
-        State.destroy({
-          uuid: json.uuid
-        }).then(function(state) {
-          ch.sendToQueue(msg.properties.replyTo,
-            new Buffer.from(JSON.stringify(state.responsify())),
-            { correlationId: msg.properties.correlationId });
-          ch.ack(msg);
-        }).catch(function(error) {
+        State.find({
+          where: {
+            uuid: json.uuid
+          }
+        }).then(function (state) {
+          if (state == undefined) {
+            ch.sendToQueue(msg.properties.replyTo,
+              new Buffer("Unknown state."),
+              { correlationId: msg.properties.correlationId });
+            ch.ack(msg);
+          } else if (state.level < 4) {
+            ch.sendToQueue(msg.properties.replyTo,
+              new Buffer("Can't delete this state."),
+              { correlationId: msg.properties.correlationId });
+            ch.ack(msg);
+          } else {
+            Task.find({
+              where: {
+                stateId: state.id
+              }
+            }).then(function (task) {
+              if (task != undefined) {
+                ch.sendToQueue(msg.properties.replyTo,
+                  new Buffer("Can't delete this state as long as it has tasks."),
+                  { correlationId: msg.properties.correlationId });
+                ch.ack(msg);
+              } else {
+                State.destroy({
+                  uuid: json.uuid
+                }).then(function(state) {
+                  ch.sendToQueue(msg.properties.replyTo,
+                    new Buffer.from(JSON.stringify(state.responsify())),
+                    { correlationId: msg.properties.correlationId });
+                  ch.ack(msg);
+                }).catch(function(error) {
+                  console.log(error);
+                  ch.sendToQueue(msg.properties.replyTo,
+                    new Buffer(error.toString()),
+                    { correlationId: msg.properties.correlationId });
+                  ch.ack(msg);
+                });
+              }
+            }).catch(function (error) {
+              console.log(error);
+              ch.sendToQueue(msg.properties.replyTo,
+                new Buffer(error.toString()),
+                { correlationId: msg.properties.correlationId });
+              ch.ack(msg);
+            });
+          }
+        }).catch(function (error) {
           console.log(error);
           ch.sendToQueue(msg.properties.replyTo,
             new Buffer(error.toString()),
